Bumpless Lock to Unlock Transition for DAC 1
This had been multiplied by 2**4, a holdover from the original NIST Digital Servo project when this DAC channel was 20 bits. The readings from DAC 1 should be the same size as DAC 0. With this extra factor, the DAC offset would jump to a position away from the last lock point, which is not the intended action.