Send Counter Mode to RP on Init
This matches the counter mode settings between sl and the frequency error widget and sends that value to the red pitaya on initialization. Since there are differences in how to correctly convert counter values read from the FPGA into Hz based on the counter type, this ensures that the calculated counts are accurate. For example, if the red pitaya was left on rectangular counting from a previous instance of the UI, and the default in the code was triangular, the new counter values would be off by 8 orders of magnitude after initialization without this fix (see sl.scaleCounterReadingsIntoHz).