Reconnection to RP
Each widget have a function to read from the RP ( XEM_GUI_MainWindow, LoopFiltersUI, FreqErrorWindowWithTempControlV2, DisplayDividerAndResidualsStreamingSettingsWindow, DisplayDitherSettingsWindow, ConfigurationRPSettingsUI)

At the opening, you have the choice to push the default values (from the xml file) to the RP, to reconnect to the RP or to open the GUI without any RP.

XEM_GUI3 have the shape of a class, with new pushValues, getValues and close connection methods.

In RP_PLL, there is a timeout of 2sec for the communication. If there is no answer from the RP before the timeout, an exception is raised and the communication is ended (the GUI does not freeze anymore)

Most of the class now take the controller ( XEM_GUI3) as a parameter to call the communication and connection methods.

In the settings tab, you can open the communication menu (the same from the opening of the program) to reconnect to a Red Pitaya.