Add option to change host port
Add option to change the host port. Usefull to connect remotly to a router with red pitaya.

You have to use the manual entry to enter the IP address of the router and the port to which the RP is redirected in the virtual server.

Redirect the private port 5000 of the Red Pitaya's IP address to the port you want to connect to. Then enter this port number in the gui beside the IP address of the router.