Uptime Logs
Records the duration of the auto recoverty session and any lost lock
events. Times are given as POSIX time stamps (UTC), which label the
beginning and ending of an event. Data is saved in HDF5 format
(PyTables), with the first row being the total session time and all
others the unlock times. New tables are createrd in the same .h5 file
upon the start of each new auto recovery session. New files are only
created at the first initialization of the UI.