Speed Up DAC Plots
The DAC plots now update with each new datapoint, on each loop. Changed
the plot buffer arrays so that there can be different numbers of counter
and DAC points. Decreased the loop timer to 200ms from 500ms to increase
the DAC sample rate. Restructured the displayFreqCounter script so that a
valid timestamp is needed, as plot points are now dropped by age.
Changed the text formatting on the display to keep the numbers more
reader friendly. The SI_scale function returns a formatted string such
that the number has a power of 10 with an exponent that is a multiple of 3, and contains the desired amount of significant digits (min 3).