User Input Error Handling
QLineEdits now have a QDoubleValidator so that the program does not try
to read invalid input. The other major addition is that values are only
read from the UI after editingFinished is called, so incomplete edits
are ignored. Implemented these changes in "user_friendly_QLineEdit" and
then used the user friendly version to create all of the line edits on the
counters tab. Also made some small changes to the layout.