Limit DAC Output Range
Added an option on the UI which limits the output range of the DAC to
just outside the auto recovery threshold. Added a
"restricted_DACs_limit" variabled to SuperLaserLand, calling
"set_dac_limits" updates this variable for use in the rest of the UI.
The "DACs_limit" variables are unchanged and serve as a record of the
full DAC ouput range. The limit is only active if the checkbox is
enabled and it is currently an auto recovery session (both lock and auto
recovery enabled), otherwise the DACs revert back to the full output
limits. The recovery thresholds are plotted as dotted lines and the DAC
limits are plotted as dashed lines. Also cleaned up some the plotting
and printing code.