Auto Recovery Baseline
From Connor Fredrick: This adds the core of the auto recovery script, which relocks the servo
at its running average if the current DAC output is outside of a set
number of standard deviations. Replaces the frequency control loop and
adds relevant controls to the GUI. Requires access to the
xem_gui_mainwindow as passed in 'XEM_GUI3.py'.